pyproject.toml
[tool.poetry]
name = "my_Proj"
version = "0.1.0"
description = ""
authors = ["Elena"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
poetry-core = "^1.9.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
types-requests = "^2.32.0.20241016"
stubs = "^1.0.0"
mock = "^5.1.0"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = '.venv'


[tool.black]
line_length = 119
exclude ='.git'


[tool.isort]
line_length = 119
